skill_training/requirements.txt
streamlit>=1.44.0
Pillow>=11.1.0 